Update calculator.py
Using `func_name` in line 96, and not defining it in line 92 was causing an unexpected `NameError`, which was being picked up by the code in `application` and resulted in a 404. This is kind of an unfortunate issue, but show up in office hours if you'd like more help addressing it!